Add authentication to quiz sessions API and create missing route files
The quiz session route had no authentication checks and exported
submitAnswer and submitQuiz functions that were unreachable via
HTTP. The client-side code in quiz-session-manager.js called
endpoints like /api/quiz-sessions/create, /api/quiz-sessions/answer,
and /api/quiz-sessions/submit that did not exist as route files.

Split the monolithic route.js into a proper Next.js App Router
structure with each endpoint in its own route file, added requireAuth
calls to every handler, and verified ownership of each session
before allowing access.
