+++ app/api/quiz-sessions/create/route.js
@@ -1,0 +1,1 @@
+export { POST } from "../route";
